create Query.join_with/3
create a query function that allows creating a named join with a dynamic
alias, Ecto.Query.join/5 does not allow that and requires the atom to be
passed at compile time which is not viable through a regular function if
we want to allow the API client to pass a custom alias.
we need the a named join in the list players query so we can reference
the player_statistic inside the query to sort the results.